% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fair_representation.R
\name{Fair_pie}
\alias{Fair_pie}
\title{Fair_pie
Make a pie chart describing the results}
\usage{
Fair_pie(Suite_results)
}
\arguments{
\item{Suite_results}{results of the suite of test}
}
\value{
a pie chart
}
\description{
Fair_pie
Make a pie chart describing the results
}
