% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fair_representation.R
\name{Fair_scores}
\alias{Fair_scores}
\title{Fair_scores
Function rendering a Fair Quality Score}
\usage{
Fair_scores(Res, dir = "../checks")
}
\arguments{
\item{Res}{Result of a test suite}

\item{dir}{directories where are situated checks}
}
\value{
a barplot with scores
}
\description{
Fair_scores
Function rendering a Fair Quality Score
}
