% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fair_representation.R
\name{Fair_table}
\alias{Fair_table}
\title{Fair_Table
Make a table of messages for checks}
\usage{
Fair_table(Suite_results, dir = "../checks")
}
\arguments{
\item{Suite_results}{results of a test suite}

\item{dir}{directories where are situated checks}
}
\value{
table of checks descriptions
}
\description{
Fair_Table
Make a table of messages for checks
}
