% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_down.R
\name{map_geographical_coverage}
\alias{map_geographical_coverage}
\title{map_geographical_coverage
Make a map from EML}
\usage{
map_geographical_coverage(eml)
}
\arguments{
\item{eml}{Metadata using EML standard in XML format}
}
\value{
A map
}
\description{
map_geographical_coverage
Make a map from EML
}
