% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_down.R
\name{render_eml}
\alias{render_eml}
\title{Render EML}
\usage{
render_eml(
  file,
  open = FALSE,
  outfile = "DataPaper.html",
  publish_mode = TRUE,
  output_dir = "/docs",
  encoding = ""
)
}
\arguments{
\item{file}{A valid Ecological Metadata Language file to be rendered to html.}

\item{open}{Whether to open the file in a browser. Defaults to TRUE.}

\item{outfile}{Name of output file.}

\item{publish_mode}{TRUE. If TRUE the website is pretty without warnings for weird stuff.}

\item{output_dir}{directory where will be stored the result file}

\item{encoding}{"" encoding of the EML file if necessary}
}
\value{
HTML file containing dataset information
}
\description{
Render EML metadata into a webpage
}
\details{
Pass in an xml file of EML metadata and generate a nice webpage describing
the dataset.
}
