% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_down.R
\name{write_custom_css}
\alias{write_custom_css}
\title{write_custom_css
Write a CSS}
\usage{
write_custom_css(publish_mode)
}
\arguments{
\item{publish_mode}{Boolean}
}
\value{
a CSS file
}
\description{
write_custom_css
Write a CSS
}
